+++ src/Adapter/SQL/CollectionTable.php
@@ -34,7 +34,6 @@ final class CollectionTable
 {
     /** @var Definition<T> */
     private Definition $definition;
-    private Table\Name\Aliased $main;
     private Table\Name\Aliased $name;
     /** @var Set<Column\Name\Aliased> */
     private Set $columns;
@@ -49,7 +48,6 @@ final class CollectionTable
         Table\Name\Aliased $main,
     ) {
         $this->definition = $definition;
-        $this->main = $main;
         $this->name = Table\Name::of($main->name()->toString().'_'.$definition->name())->as($definition->name());
         $this->columns = $definition
             ->properties()